v2.0

+++ WinkAtHome/Classes/SettingMgmt.cs
@@ -1,6 +1,7 @@
 ﻿using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
+using System.Data;
 using System.Data.SQLite;
 using System.Diagnostics;
 using System.IO;
@@ -13,7 +14,15 @@ namespace WinkAtHome
     public class SettingMgmt
     {
         private static string dbPath = Common.dbPath;
-        
+
+        public class stringbool : Tuple<string, bool>
+        {
+            public stringbool(string value, bool isEncrypted)
+                : base(value, isEncrypted)
+            {
+            }
+        }
+
         public class Setting
         {
             public string key;
@@ -46,8 +55,8 @@ namespace WinkAtHome
             }
             catch (Exception ex)
             {
-                throw ex; //EventLog.WriteEntry("WinkAtHome.SettingsMgmt.getSetting", ex.Message, EventLogEntryType.Error);
                 return null;
+                throw ex; //EventLog.WriteEntry("WinkAtHome.SettingsMgmt.getSetting", ex.Message, EventLogEntryType.Error);
             }
         }
 
@@ -61,35 +70,90 @@ namespace WinkAtHome
 
                     using (SQLiteConnection connection = new SQLiteConnection("Data Source=" + dbPath + ";Version=3;"))
                     {
-                        string sql = "select * from Settings";
-
                         connection.Open();
-                        SQLiteCommand command = new SQLiteCommand(sql, connection);
-                        SQLiteDataReader reader = command.ExecuteReader();
-                        while (reader.Read())
+
+                        //LEGACY CORRECTIONS
+                        using (SQLiteCommand command = new SQLiteCommand(connection))
                         {
-                            Setting setting = new Setting();
-                            setting.key = reader["name"].ToString();
-                            string isEncrypted = reader["IsEncrypted"].ToString();
-                            string isRequired = reader["isRequired"].ToString();
+                            command.CommandText = "DELETE FROM Settings WHERE Name='winkPassword' or Name='winkUsername'";
+                            command.ExecuteNonQuery();
 
-                            if (reader.IsDBNull(reader.GetOrdinal("value")) || string.IsNullOrWhiteSpace(reader["value"].ToString()))
-                                setting.value = reader["defaultvalue"].ToString();
-                            else
+                            if (Wink.winkUser != null)
                             {
-                                if (Convert.ToBoolean(isEncrypted))
-                                    setting.value = Common.Decrypt(reader["value"].ToString());
-                                else
-                                    setting.value = reader["value"].ToString();
+                                command.CommandText = "UPDATE Settings SET UserID=@UserID WHERE UserID='single'";
+                                command.Parameters.Add(new SQLiteParameter("@UserID", Wink.winkUser.userID));
+                                command.ExecuteNonQuery();
                             }
+                        }
 
-                            if (Convert.ToBoolean(isEncrypted))
-                                setting.isEncrypted = true;
-                            
-                            if (Convert.ToBoolean(isRequired))
-                                setting.isRequired = true;
+                        //INSERT DEFAULT SETTINGS
+                        Dictionary<string, stringbool> basicSettings = new Dictionary<string, stringbool>();
+                        basicSettings.Add("winkClientID", (new stringbool("quirky_wink_android_app", true)));
+                        basicSettings.Add("winkClientSecret", (new stringbool("e749124ad386a5a35c0ab554a4f2c045", true)));
+                        basicSettings.Add("PubNub-PublishKey", (new stringbool(null, true)));
+                        basicSettings.Add("PubNub-SubscribeKey", (new stringbool(null, true)));
+                        basicSettings.Add("PubNub-SecretKey", (new stringbool(null, true)));
+                        basicSettings.Add("StartPage", (new stringbool("Control.aspx", false)));
+                        basicSettings.Add("Hide-Empty-Robots", (new stringbool("false", false)));
+                        basicSettings.Add("Hide-Empty-Groups", (new stringbool("false", false)));
+                        basicSettings.Add("Robot-Alert-Minutes-Since-Last-Trigger", (new stringbool("60", false)));
 
-                            settings.Add(setting);
+                        foreach (KeyValuePair<string, stringbool> pair in basicSettings)
+                        {
+                            using (SQLiteCommand defaultcommand = new SQLiteCommand(connection))
+                            {
+                                defaultcommand.CommandText = "UPDATE Settings SET DefaultValue=@defaultvalue, IsEncrypted=@isEncypted, IsRequired=@isRequired WHERE name = @name;";
+                                defaultcommand.Parameters.Add(new SQLiteParameter("@name", pair.Key));
+                                defaultcommand.Parameters.Add(new SQLiteParameter("@defaultvalue", pair.Value.Item1));
+                                defaultcommand.Parameters.Add(new SQLiteParameter("@isEncypted", pair.Value.Item2));
+                                defaultcommand.Parameters.Add(new SQLiteParameter("@isRequired", true));
+                                defaultcommand.ExecuteNonQuery();
+
+                                defaultcommand.CommandText = "SELECT DISTINCT UserID FROM Users;";
+                                SQLiteDataAdapter da = new SQLiteDataAdapter(defaultcommand);
+                                DataTable dtTable = new DataTable();
+                                da.Fill(dtTable);
+
+                                foreach (DataRow row in dtTable.Rows)
+                                {
+                                    defaultcommand.CommandText = "INSERT OR IGNORE INTO Settings(UserID,Name,DefaultValue,IsEncrypted,IsRequired) VALUES ('" + row[0] + "',@name,@defaultvalue,@isEncypted,@isRequired)";
+                                    defaultcommand.ExecuteNonQuery();
+                                }
+                            }
+                        }
+
+                        using (SQLiteCommand command = new SQLiteCommand(connection))
+                        {
+                            //PROCESS SETTINGS
+                            command.CommandText = "select * from Settings WHERE UserID = @UserID";
+                            command.Parameters.Add(new SQLiteParameter("@UserID", Wink.winkUser.userID));
+
+                            SQLiteDataReader reader = command.ExecuteReader();
+                            while (reader.Read())
+                            {
+                                Setting setting = new Setting();
+                                setting.key = reader["name"].ToString();
+                                string isEncrypted = reader["IsEncrypted"].ToString();
+                                string isRequired = reader["isRequired"].ToString();
+
+                                if (reader.IsDBNull(reader.GetOrdinal("value")) || string.IsNullOrWhiteSpace(reader["value"].ToString()))
+                                    setting.value = reader["defaultvalue"].ToString();
+                                else
+                                {
+                                    if (Convert.ToBoolean(isEncrypted))
+                                        setting.value = Common.Decrypt(reader["value"].ToString());
+                                    else
+                                        setting.value = reader["value"].ToString();
+                                }
+
+                                if (Convert.ToBoolean(isEncrypted))
+                                    setting.isEncrypted = true;
+
+                                if (Convert.ToBoolean(isRequired))
+                                    setting.isRequired = true;
+
+                                settings.Add(setting);
+                            }
                         }
                     }
 
@@ -100,7 +164,6 @@ namespace WinkAtHome
             catch (Exception ex)
             {
                 throw ex; //EventLog.WriteEntry("WinkAtHome.SettingsMgmt.loadSettings", ex.Message, EventLogEntryType.Error);
-                throw ex;
             }
         }
 
@@ -116,15 +179,16 @@ namespace WinkAtHome
                     connection.Open();
                     using (SQLiteCommand command = new SQLiteCommand(connection))
                     {
-                        command.CommandText = "UPDATE Settings SET Value=@value WHERE name = @name;";
+                        command.CommandText = "UPDATE Settings SET Value=@value WHERE UserID = @UserID AND name = @name;";
+                        command.Parameters.Add(new SQLiteParameter("@UserID", Wink.winkUser.userID));
                         command.Parameters.Add(new SQLiteParameter("@name", key));
                         command.Parameters.Add(new SQLiteParameter("@value", newValue));
                         command.ExecuteNonQuery();
 
-                        command.CommandText = "INSERT OR IGNORE INTO Settings(Name,Value) VALUES (@name,@value)";
+                        command.CommandText = "INSERT OR IGNORE INTO Settings(UserID, Name,Value) VALUES (@UserID, @name,@value)";
                         command.ExecuteNonQuery();
 
-                        command.CommandText = "DELETE FROM Settings WHERE IFNULL(Value, '') = '' and IFNULL(DefaultValue, '') = '' and IsEncrypted = 'false'";
+                        command.CommandText = "DELETE FROM Settings WHERE UserID = @UserID AND IFNULL(Value, '') = '' and IFNULL(DefaultValue, '') = '' and IsEncrypted = 'false'";
                         command.ExecuteNonQuery();
                     }
                 }
@@ -165,10 +229,11 @@ namespace WinkAtHome
 
                     using (SQLiteCommand command = new SQLiteCommand(connection))
                     {
-                        command.CommandText = "delete from Settings where defaultvalue is null";
+                        command.CommandText = "delete from Settings where UserID = @UserID AND defaultvalue is null";
+                        command.Parameters.Add(new SQLiteParameter("@UserID", Wink.winkUser.userID));
                         command.ExecuteNonQuery();
 
-                        command.CommandText = "update Settings set value = null";
+                        command.CommandText = "update Settings set value = null WHERE UserID = @UserID";
                         command.ExecuteNonQuery();
 
                     }
@@ -180,7 +245,6 @@ namespace WinkAtHome
             {
                 throw ex; //EventLog.WriteEntry("WinkAtHome.SettingsMgmt.wipeSettings", ex.Message, EventLogEntryType.Error);
             }
-
         }
     }
 }
